+++ src/components/radio/mdc-radio.js
@@ -1,6 +1,5 @@
 import { inject, bindable, bindingMode, DOM } from 'aurelia-framework';
-import { radio as mdcRadio } from 'material-components-web/dist/material-components-web';
-const { MDCRadio } = mdcRadio;
+import { MDCRadio } from '@material/radio/dist/mdc.radio.min';
 
 @inject(Element)
 export class MdcRadio {